fix stringutil long parse

+++ src/main/java/com/go2wheel/mysqlbackup/util/StringUtil.java
@@ -52,7 +52,7 @@ public class StringUtil {
 	public static long parseLong(String numberHeaded) {
 		Matcher m = NUMBER_HEADED.matcher(numberHeaded);
 		if (m.matches()) {
-			return Integer.valueOf(m.group(1));
+			return Long.valueOf(m.group(1));
 		} else {
 			return 0;
 		}
